CRA-221: Removed check of surface covering, and added some well interpolation when new keyword <rms-panel-mode> is used.
Also removed old model.h and model.cpp, see previous commit by Heidi.


git-svn-id: https://code.nr.no/scm/svn/crava/trunk@1180 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/analyzelog.cpp
@@ -594,7 +594,7 @@ Analyzelog::estimateCorrTAndVar0(float       * CorrT,
       Var0[1][1] = 2*Var0[0][0];
     }
     // Default covariance between Vp and Vs. Value is explained in Jira issue CRA-220.
-    Var0[0][1] = Var0[1][0] = std::sqrt(Var0[0][0])*std::sqrt(Var0[1][1])*0.70; 
+    Var0[0][1] = Var0[1][0] = static_cast<float>(std::sqrt(Var0[0][0])*std::sqrt(Var0[1][1])*0.70); 
     Var0[0][2] = Var0[2][0] = 0.0;
     Var0[1][2] = Var0[2][1] = 0.0;
     if (nAB>1)
